split ExpanderWidget into its own set of files
