+++ src/modinfodialog.h
@@ -27,6 +27,7 @@ along with Mod Organizer.  If not, see <http://www.gnu.org/licenses/>.
 #include "organizercore.h"
 #include "filterwidget.h"
 #include "filerenamer.h"
+#include "expanderwidget.h"
 
 #include <QDialog>
 #include <QSignalMapper>
@@ -51,47 +52,6 @@ class QFileSystemModel;
 class QTreeView;
 class CategoryFactory;
 class TextEditor;
-
-
-/* Takes a QToolButton and a widget and creates an expandable widget.
- **/
-class ExpanderWidget
-{
-public:
-  /** empty expander, use set()
-   **/
-  ExpanderWidget();
-
-  /** see set()
-   **/
-  ExpanderWidget(QToolButton* button, QWidget* content);
-
-  /** @brief sets the button and content widgets to use
-    * the button will be given an arrow icon, clicking it will toggle the
-    * visibility of the given widget
-    * @param button the button that toggles the content
-    * @param content the widget that will be shown or hidden
-    * @param opened initial state, defaults to closed
-    **/
-  void set(QToolButton* button, QWidget* content, bool opened=false);
-
-  /** either opens or closes the expander depending on the current state
-   **/
-  void toggle();
-
-  /** sets the current state of the expander
-   **/
-  void toggle(bool b);
-
-  /** returns whether the expander is currently opened
-   **/
-  bool opened() const;
-
-private:
-  QToolButton* m_button;
-  QWidget* m_content;
-  bool opened_;
-};
 
 
 /**
